+++ src/server/utils.ts
@@ -1,4 +1,4 @@
-import { existsSync, readFileSync } from "fs";
+import { existsSync, readFileSync, promises } from "fs";
 import { dirname, extname, join } from "path";
 import { getHash } from "@arnaud-barre/config-loader";
 
@@ -13,6 +13,7 @@ export const getHashedUrl = (base: string, content: string | Buffer) =>
   `/${base}?h=${getHash(content).slice(0, 8)}`;
 
 export const readFile = (path: string) => readFileSync(path, "utf-8");
+export const readFileAsync = (path: string) => promises.readFile(path, "utf-8");
 
 export const cacheDir = "node_modules/.rds";
 export const readCacheFile = (path: string) => readFile(join(cacheDir, path));