+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.3
+
+- Fix SWC errors display
+- Fix crash on downwind error
+- Fix SWC cache performances
+
 ## 0.1.2
 
 - Fix handling of syntax error on JS HMR update